Add specific System point with portal count
Portal count will be useful to know how much far to scan in the grabpass later.
It avoids having to pass the value to the shader directly.
It is guarded by the sentinel value to avoid garbage counts.
It also provides a vertex for the sentinel instead of adding the pixel to primitive_id 0.