Using uint4 CRT texture works nicely for roundtrip
Fixed encoding bug by using a raw uint4 texture.
F32 bit encoded as u32 are ok, and so are the pair of f16.
Using f32 as texture values adds unpredictable encoding rules that should be avoided if not everuything is f32.
Bonus point : asuint/asfloat are reinterpret casts with no cost, they do not appear in the assembly.

We loose the debug value of displayed texture colors, but this can be fixed by debug shaders.
Next : convert everything to CRT format, and then update grabpass format.