Convert to new encoding for grabpass
Use the same portal struct with 2 sets of encoding functions.
Grabpass encoding now encodes a full f32x3 for world position, using bit encoding.
This costs 4 pixels but this is mitigated by CRT pass that compress it to 2 pixels.

Cleaned up encoding-decoding code.
Moved to one header containing both for simplicity.
Added helpers that sample the textures.

During the rework I tried to increase precision on sentinel value test, but this was a mistake and made the system fail the check.
Do not touch the threshold again !

Next : camera encoding and CRT logic.