Fixed camera state, working in desktop nicely.
Global variables for camera positions are available in CRT.
Thus ditch grabpass encodings that are useless.

Fixed visualisation to not display the camera cross if too close.
Before it would be annoying on the current camera.

Position of the CRT debug meshes seems laggy, will have to see in VR with reprojection.
Maybe mess with update zone swap settings ?