Occlusion vertices on portals instead of huge box
Generate vertices along bounding box of portal quads.
These will force proper occlusion box update, without forcing a huge box.

Will this be ok for camera state ?
I guess so, as the CRT would update when the first portal is seen by a camera, which may be the photo one if faraway.