Fix near plane clipping on portals
When in world mode, portals are used to enable pixel test on part of the screen.
The actual depth of the portal is irrelevant as depth will be computed manually.
Thus near plane clipping is bad as it will remove some pixels, especially when traversing a portal (very close to it).
Thus tweak the view space position (by scaling xyz) so that screenspace positions are kept, but the quad is away from near plane.
Strategy is to scale so that the camera distance to quad plane is more than corner of near plane frustum (+margin).

Desktop test : works.
There are still problems that after tests are due to camera pos in CRT being from the last frame (like portal positions).
TODO fix.