WIP Mesh probe automatic definition from mesh
Manual definition of mesh probe is way too annoying / tedious to use in practice.
Instead pivot to an automatic + hint system that look at vertex distributions.

Concept for now :
- scan all renderer meshes, and create probes for each bone : done except basic renderer which will likely require manual hints to be useful
- hints to merge children probes into self (useful for hands) : done
- set probe index in mesh uvs : done
- generate radius from probe vertices TODO
- manual placement hints TODO
- probe tree links setup and pixel generation TODO

To avoid long computation time at upload, I try to keep everything single pass on the meshes.
For now the idea is to do a single scan on each mesh vertice.
Probe position + radius is intended to be barycenter + distance, and requires 2 pass ; for now collect list of local vertice positions to be used later.
For now each vertice gets up to 2 probe indexes in uvs, x being the stronger one. Not sure if it will be used later but this is no more expensive.

Probes are only created for a bone if an influenced vertex actually exists. This eliminates spurious bones that are included but never inflence (`_end`, intermediate, etc).
On the test avatar :
- 130 skinned weights by default
- merging fingers gets us to 90
- adding lazy bone probe creation gets to 80, at no visible time cost. But sample avatar has a low triangle count (10k, 30ms for now), keep monitoring this.