Fix pixel tag states : forgot portal space without intersection
Sealing shader was behaving badly for portal space object without intersection when camera in portal.
Fix by extending the cases for pixel tags. Now 3 states !

Last commit was tested in editor only, not in game where camera state can be checked.
This commit is checked ingame and should be final for the logic. Now it is time to integrate into shaders.