Prototype integration into unity standard shader
VRChat mobile shaders are surface shaders, so not a good target due to their outdated support.
Using Unity Standard, with as few modifications as possible.

Only opaque supported for now.

Added an integration method specific to shadowcaster.
After slight testing, the most effective semantic seems to be to just cut fragment pixels in portal space.

TODO HeadChop fix before a full prototype can be validated.
But in the meantime this is the first minimum viable system.