Improve precision of mesh probes
Mesh probe radius was not precise and part of the mesh would show through the other side when traversing.

Added a margin slider to add padding to the computed radius.
Added secondary bone influences to the list of probe vertices, so that radius computation takes them into account.

The better solution would be to look at triangles to add linked vertices to a probe, but more costly.
This will wait for a mesh probe v2 with overrides and/or clustering.

Also removed merger from one hand for tests.
Mesh probe count is not that problematic so maybe they are not that useful anyway...