Main "discard bug" fixed
Discard bug was a shader miscompilation error.
In SPS-I variant, the compiler generates wrong code when retrieving the stereo camera position in the specific branch case of the bug.
Camera position is retrieved by dynamically indexing in the constant array with stereoIndex.
With the miscompile, the index is not initialized in this branch, and this retrieves garbage data, and the depth computation is bogus.

Reorganizing code to mutualise discard at the end did not change anything.
Precomputing the camera pos + runtime 0 to prevent inlining works, but I had to dance around compiler internal errors.
This is worrying as more code needs to be added for visuals...
Maybe converting dynamic indexing in SPS-I to cmovs with stereoIndex could remove the need for dynamic indexing and help the compiler ?

There is still the lag bug in the same configuration, which is not solved not understood.