Discard bug : use cmov instead of +runtime_0
Remove the runtime 0 parameter and do a cmov on stereEyeIndex between the two array cells in SPS-I.
Compiles correctly and less variable clutter.

Additionnally it gives the option to convert other dynamically indexed constants later if needed.
Changing this one does not fix the internal compiler crash if I create an else branch, which happened in the no keyword variant anyway...