Fix incoherent mesh & main camera portal states
Camera and mesh probe states are tracked independently.
So they can be incoherent : camera in portal and mesh outside.
This is normal for remotes, but bad for local.

Thus add a IsLocal check that resets main camera state to head bone mesh probe in case of incoherence.
Mesh state is taken as more important than camera to closely match what remotes are viewing (they only see a similar mesh state as local).

Incoherence is defined as different states + not explained by a portal intersection.
So add an update intersection test head-camera.
Nice trick : the thread containing head mesh probe will be the one to compute and emit the final updated header, as it already has the updated head mesh probe state.

This upgrade is bundled with cleanup of update shader.
Header variables have been renamed to be more readable (main/photo camera instead of index).
Pack as many intersection tests in as few loops as possible :
- every thread now computes camera intersections while building portal masks. Only 1 will emit the updated header but all could do it, almost.
- reuse probe update to retrieve the new state for head bone and compute intersections.

Math cost is similar with only two portal loops now.