Fix general lag bug : it was camera order
After checking update count ("tick tock") between rt0 and rt1, it appeared that :
- it was not +1 as expected when lag is present
- but overall tick+tock increment was still +2 and state was not broken
Thus this means that update order was reversed as rt1 then rt0.

Toggling the whole system fixes it reliably on my machine ; maybe because both cameras are toggled recursively at once.
I did not remeber this being the case, but I did not specifically check this so maybe it worked like that before.
Only on init are cameras not toggled in the correct order.
The fix is adding a "camera container object" and toggling it 1 frame after cameras are enabled.

Test desktop+VR : fix is good.
It does not fix the discard lag bug at all. Sad, I still have no idea what is happening there...
It does not fix the head incoherence system having edge cases.

I though at some point of replacing with CRT to have consistent 1 frame delay, and deal with it the expensive way.
This is still on the table...