Disable mirror support entirely
It seems that mirrors are part of the discard lag bug problem.
Simple solutions like not writing weird alphas to mirror are not sufficient.

Mirror support is broken for now, so in the meantime just disable it properly and behave as if there was no portal system in the mirrors.

Anti-vampire that is visible in the mirror but not reality ?
A fun possibility would be to use mirrors as portals... probably not a good idea though.